+++ serving-core/src/test/scala/com/stratio/sparta/serving/core/actor/SchedulerMonitorActorTest.scala
@@ -185,7 +185,7 @@ class SchedulerMonitorActorTest extends TestKit(ActorSystem("SchedulerActorSpec"
       EdgeGraph("a", "b")
     )
     val settingsModel = Settings(
-      GlobalSettings(WorkflowExecutionMode.marathon, Seq.empty, Seq.empty, true,
+      GlobalSettings(WorkflowExecutionMode.marathon, Seq.empty, Seq.empty, Seq.empty, true,
         Some(JsoneyString("constraint1:constraint2"))),
       StreamingSettings(
         JsoneyString("6s"), None, None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),